feat(namadillo): small refactor on AccountOverview component
