+++ apps/namadillo/src/App/AccountOverview/AccountOverview.tsx
@@ -1,43 +1,27 @@
-import {
-  ActionButton,
-  Heading,
-  SkeletonLoading,
-  Stack,
-} from "@namada/components";
+import { Stack } from "@namada/components";
 import { useUntilIntegrationAttached } from "@namada/integrations";
-import { FiatCurrency } from "App/Common/FiatCurrency";
 import { Intro } from "App/Common/Intro";
-import { NamCurrency } from "App/Common/NamCurrency";
 import { PageWithSidebar } from "App/Common/PageWithSidebar";
-import GovernanceRoutes from "App/Governance/routes";
 import MainnetRoadmap from "App/Sidebars/MainnetRoadmap";
-import StakingRoutes from "App/Staking/routes";
 import clsx from "clsx";
 import { useAtomValue } from "jotai";
-import { useNavigate } from "react-router-dom";
-import { accountBalanceAtom } from "slices/accounts";
 import { chainAtom } from "slices/chain";
 import { namadaExtensionConnectedAtom } from "slices/settings";
+import { AccountBalanceContainer } from "./AccountBalanceContainer";
+import { NavigationFooter } from "./NavigationFooter";
 
 export const AccountOverview = (): JSX.Element => {
-  const navigate = useNavigate();
   const chain = useAtomValue(chainAtom);
   const isConnected = useAtomValue(namadaExtensionConnectedAtom);
   const extensionAttachStatus = useUntilIntegrationAttached(chain);
+  const fullWidthClassName = clsx({ "col-span-2": !isConnected });
+
+  // TODO: we need a better way to verify if extension is connected
   const currentExtensionAttachStatus =
     extensionAttachStatus[chain.extension.id];
-
-  const {
-    data: totalBalance,
-    isSuccess: balanceHasLoaded,
-    isLoading: balanceIsLoading,
-  } = useAtomValue(accountBalanceAtom);
-
   const hasExtensionInstalled =
     currentExtensionAttachStatus === "attached" ||
     currentExtensionAttachStatus === "pending";
-
-  const fullWidthClassName = clsx({ "col-span-2": !isConnected });
 
   return (
     <PageWithSidebar footerClassName={fullWidthClassName}>
@@ -55,67 +39,10 @@ export const AccountOverview = (): JSX.Element => {
             />
           </div>
         )}
-
         {isConnected && (
           <Stack gap={5} className="my-auto min-w-[365px] mx-auto py-12">
-            {balanceHasLoaded && totalBalance && (
-              <div
-                className={clsx(
-                  "relative flex flex-col leading-tight w-full aspect-square rounded-full border-[27px] border-yellow",
-                  "items-center justify-center text-center text-neutral-400"
-                )}
-              >
-                <Heading level="h3" className="text-xl neutral-600">
-                  NAM Balance
-                </Heading>
-                <NamCurrency
-                  amount={totalBalance}
-                  className="text-5xl text-white font-medium"
-                  currencySignClassName="text-xl ml-2"
-                />
-                <FiatCurrency
-                  amountInNam={totalBalance}
-                  className="text-xl font-medium"
-                />
-              </div>
-            )}
-            {balanceIsLoading && (
-              <SkeletonLoading
-                width="100%"
-                height="auto"
-                className="aspect-square mx-auto rounded-full bg-transparent border-[27px] border-neutral-800"
-              />
-            )}
-            <footer className="text-center">
-              <Heading
-                level="h3"
-                className="uppercase text-4xl text-yellow mb-2"
-              >
-                Stake & Vote Now
-              </Heading>
-              <Stack gap={3} direction="horizontal">
-                <ActionButton
-                  className="border-yellow border uppercase"
-                  onClick={() => navigate(StakingRoutes.overview().url)}
-                  size="sm"
-                  color="primary"
-                  borderRadius="sm"
-                >
-                  Stake
-                </ActionButton>
-                <ActionButton
-                  size="sm"
-                  onClick={() => navigate(GovernanceRoutes.index())}
-                  color="primary"
-                  outlined
-                  className="uppercase hover:text-rblack before:border before:border-yellow"
-                  borderRadius="sm"
-                  hoverColor="primary"
-                >
-                  Governance
-                </ActionButton>
-              </Stack>
-            </footer>
+            <AccountBalanceContainer />
+            <NavigationFooter />
           </Stack>
         )}
       </div>
