+++ apps/namadillo/src/utils/animations.ts
@@ -1,0 +1,4 @@
+export const easings = {
+  expoOut: [0.16, 1, 0.3, 1],
+  quartOut: [0.25, 1, 0.5, 1],
+};
